langpz3/requirements.txt
# Core dependencies
langgraph>=0.1.0
langchain>=0.3.0
langchain-openai>=0.3.12
langsmith>=0.3.0
pydantic>=2.0,<3.0

# Additional dependencies
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
multidict==6.0.4
langchain-cli==0.0.36
langgraph-cli==0.1.61
